Implement pattern-based inclusion/exclusions (#46)
This implements the ability to include/exclude files from
stripping.

Each filename that the plugin inspects for stripping is first passed
to a series of inclusion filters, and then to a series of exclusion
filters. The filters are expressed as standard java.util.Pattern
regular expressions. If the filename matches at least one inclusion
filter, then it becomes a candidate for stripping. If the candidate
filename matches at least one exclusion filter, then the file is no
longer a candidate for stripping. By default, the plugin behaves as
if no exclusion filter was provided, and exactly one inclusion filter
was provided that matches all input (.*).

Fix: https://github.com/Zlika/reproducible-build-maven-plugin/issues/36